apps/worker/pyproject.toml
[project]
name = "workey-worker"
version = "0.1.0"
description = "Workey background worker - job scraping, email monitoring, scheduled tasks"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
